moved fixtures to conftest
